@@ -13,8 +13,8 @@
   "author": "Graylog, Inc. <hello@graylog.com>",
   "license": "MIT",
   "dependencies": {
-    "eslint": "5.16.0",
-    "eslint-config-airbnb": "17.1.1",
+    "eslint": "6.8.0",
+    "eslint-config-airbnb": "18.1.0",
     "eslint-import-resolver-webpack": "0.12.1",
     "eslint-plugin-flowtype": "^4.5.2",
     "eslint-plugin-import": "2.20.1",
