@@ -34,25 +34,25 @@
     "README.md"
   ],
   "dependencies": {
-    "@babel/preset-env": "7.8.7",
+    "@babel/preset-env": "7.9.0",
     "babel-eslint": "9.0.0",
     "eslint-config-graylog": "file:../eslint-config-graylog",
-    "html-webpack-plugin": "3.2.0",
+    "html-webpack-plugin": "4.0.2",
     "javascript-natural-sort": "0.7.1",
     "jquery": "3.4.1",
     "moment": "2.24.0",
     "moment-timezone": "0.5.28",
     "prop-types": "15.7.2",
-    "react": "16.12.0",
+    "react": "16.13.1",
     "react-addons-pure-render-mixin": "15.6.2",
     "react-bootstrap": "0.33.1",
-    "react-dom": "16.12.0",
+    "react-dom": "16.13.1",
     "react-router": "3.2.1",
     "react-router-bootstrap": "0.23.2",
     "reflux": "0.2.13",
-    "styled-components": "4.3.2",
+    "styled-components": "4.4.1",
     "styled-theming": "2.2.0",
-    "webpack": "4.41.6",
+    "webpack": "4.42.1",
     "webpack-cleanup-plugin": "0.5.1",
     "webpack-cli": "3.3.11",
     "webpack-merge": "4.2.2"
