@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const logoUrl = require('images/toplogo.svg');
+import logoUrl from 'images/toplogo.svg';
 
 // Don't pass active prop, since `a` tag doesn't support it.
 // eslint-disable-next-line no-unused-vars
