@@ -5,7 +5,7 @@ import { Pagination as BootstrapPagination } from 'react-bootstrap';
 import { Pagination as DeprecatedPagination } from '@react-bootstrap/pagination';
 import { css } from 'styled-components';
 
-import { DEPRECATION_NOTICE } from 'util/constants';
+import deprecationNotice from 'util/deprecationNotice';
 import { util } from 'theme';
 
 export const paginationStyles = css(({ theme }) => css`
@@ -64,8 +64,7 @@ const Pagination = ({
 }) => {
   if (activePage || maxButtons || prev || next || first || last) {
     useEffect(() => {
-      /* eslint-disable-next-line no-console */
-      console.warn(DEPRECATION_NOTICE, 'You have used a deprecated `Pagination` prop, please check the documentation to use the latest `Pagination`.');
+      deprecationNotice('You have used a deprecated `Pagination` prop, please check the documentation to use the latest `Pagination`.');
     }, []);
 
     return (
