@@ -11,7 +11,7 @@ import Routes from 'routing/Routes';
 const HelpMenu = ({ active }) => (
   <NavDropdown active={active}
                id="help-menu-dropdown"
-               title={<Icon name="question-circle" aria-label="Help" />}
+               title={<Icon name="question-circle" size="lg" aria-label="Help" />}
                noCaret>
 
     <LinkContainer to={Routes.getting_started(true)}>
