@@ -7,21 +7,21 @@ describe('contrastingColor', () => {
     const color3 = contrastingColor('#f0f');
     const color4 = contrastingColor('#000', 'AA');
 
-    expect(color1).toBe('#999');
-    expect(color2).toBe('#595959');
-    expect(color3).toBe('#fff2ff');
-    expect(color4).toBe('#7f7f7f');
+    expect(color1).toBe('rgb(151,151,151)');
+    expect(color2).toBe('rgb(81,81,81)');
+    expect(color3).toBe('rgb(255,249,255)');
+    expect(color4).toBe('rgb(128,128,128)');
   });
 
   it('should accept other color strings', () => {
     const color1 = contrastingColor('rgb(0, 0, 0)');
 
-    expect(color1).toBe('#999');
+    expect(color1).toBe('rgb(151,151,151)');
   });
 
   it('should accept transparent color strings', () => {
     const color1 = contrastingColor('rgba(0, 0, 0, 0.5)');
 
-    expect(color1).toBe('rgba(157,157,157,0.825)');
+    expect(color1).toBe('rgba(151,151,151,0.675)');
   });
 });
