@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { createRef, Component } from 'react';
 import PropTypes from 'prop-types';
 import { Portal } from 'react-portal';
 import { Position } from 'react-overlays';
@@ -6,32 +6,48 @@ import { Position } from 'react-overlays';
 import { Popover } from 'components/graylog';
 import { Icon } from 'components/common';
 
-class HoverForHelp extends React.Component {
-  state = {
-    hover: false,
-  };
+class HoverForHelp extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      hover: false,
+    };
+
+    this.hoverTarget = createRef();
+  }
 
   _onToggleHover = () => this.setState(({ hover }) => ({ hover: !hover }));
 
-  render() {
-    const { children, title } = this.props;
+  _renderPopover = () => {
     const { hover } = this.state;
-    const popover = hover && (
+    const { children, title } = this.props;
+
+    if (!hover) {
+      return null;
+    }
+
+    return (
       <Portal>
         <Position container={document.body}
                   placement="bottom"
-                  target={this.target}>
+                  target={this.hoverTarget.current}>
           <Popover title={title} id="configuration-popover">
             {children}
           </Popover>
         </Position>
       </Portal>
     );
+  }
 
+  render() {
     return (
-      <span onMouseEnter={this._onToggleHover} onMouseLeave={this._onToggleHover}>
-        <Icon name="question-circle" className="pull-right" ref={(elem) => { this.target = elem; }} />
-        {popover}
+      <span onMouseEnter={this._onToggleHover}
+            onMouseLeave={this._onToggleHover}
+            ref={this.hoverTarget}
+            className="pull-right">
+        <Icon name="question-circle" />
+        {this._renderPopover()}
       </span>
     );
   }
