@@ -40,13 +40,13 @@ describe('RefreshControls', () => {
 
   it('should start the interval', () => {
     const wrapper = mount(<RefreshControls refreshConfig={{ enabled: false, interval: 1000 }} />);
-    wrapper.find('i.fa-play').simulate('click');
+    wrapper.find('svg.fa-play').simulate('click');
     expect(RefreshActions.enable).toHaveBeenCalled();
   });
 
   it('should stop the interval', () => {
     const wrapper = mount(<RefreshControls refreshConfig={{ enabled: true, interval: 1000 }} />);
-    wrapper.find('i.fa-pause').simulate('click');
+    wrapper.find('svg.fa-pause').simulate('click');
     expect(RefreshActions.disable).toHaveBeenCalled();
   });
 });
