@@ -1,4 +1,4 @@
-import AceEditor from 'react-ace-builds';
+import AceEditor from 'react-ace';
 import 'views/components/searchbar/queryinput/webpack-resolver';
 import 'views/components/searchbar/queryinput/ace-queryinput';
 import 'ace-builds/src-min-noconflict/ext-language_tools';
