@@ -12,7 +12,7 @@ import XYPlot from '../XYPlot';
 
 const seriesGenerator = (type, name, labels, values) => ({ type, name, x: labels, y: values, mode: 'markers' });
 
-const ScatterVisualization: VisualizationComponent = makeVisualization(({ config, data, effectiveTimerange }: VisualizationComponentProps) => {
+const ScatterVisualization: VisualizationComponent = makeVisualization(({ config, data, effectiveTimerange, height }: VisualizationComponentProps) => {
   const chartDataResult = chartData(config, data.chart || Object.values(data)[0], 'scatter', seriesGenerator);
   const layout = {};
   if (config.eventAnnotation && data.events) {
@@ -24,6 +24,7 @@ const ScatterVisualization: VisualizationComponent = makeVisualization(({ config
     <XYPlot config={config}
             chartData={chartDataResult}
             plotLayout={layout}
+            height={height}
             effectiveTimerange={effectiveTimerange} />
   );
 }, 'scatter');
@@ -31,6 +32,7 @@ const ScatterVisualization: VisualizationComponent = makeVisualization(({ config
 ScatterVisualization.propTypes = {
   config: AggregationType.isRequired,
   data: PropTypes.arrayOf(PropTypes.object).isRequired,
+  height: PropTypes.number,
 };
 
 export default ScatterVisualization;
