@@ -17,19 +17,19 @@
 package integration;
 
 import com.github.zafarkhaja.semver.Version;
-import com.jayway.restassured.RestAssured;
-import com.jayway.restassured.authentication.AuthenticationScheme;
-import com.jayway.restassured.builder.RequestSpecBuilder;
-import com.jayway.restassured.config.MatcherConfig;
+import io.restassured.RestAssured;
+import io.restassured.authentication.AuthenticationScheme;
+import io.restassured.builder.RequestSpecBuilder;
+import io.restassured.config.MatcherConfig;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 import org.junit.rules.ExternalResource;
 
 import java.net.URI;
 
-import static com.jayway.restassured.RestAssured.given;
-import static com.jayway.restassured.RestAssured.preemptive;
-import static com.jayway.restassured.http.ContentType.JSON;
+import static io.restassured.RestAssured.given;
+import static io.restassured.RestAssured.preemptive;
+import static io.restassured.http.ContentType.JSON;
 
 public class RestAssuredSetupRule extends ExternalResource {
     private static Version serverUnderTestVersion;
@@ -55,7 +55,7 @@ public class RestAssuredSetupRule extends ExternalResource {
 
         // immediately query the server for its version number, we need it to be able to process the RequireVersion annotations.
         given().when()
-                    .authentication().preemptive().basic(userInfo[0], userInfo[1])
+                .auth().basic(userInfo[0], userInfo[1])
                     .get("/system")
                 .then()
                     .body("version", new BaseMatcher<Object>() {
