@@ -19,11 +19,12 @@ package integration.system;
 import integration.BaseRestTest;
 import integration.RequiresAuthentication;
 import integration.RequiresVersion;
+import org.junit.Ignore;
 import org.junit.Test;
 
-import static com.jayway.restassured.RestAssured.given;
-import static org.junit.Assert.*;
+import static io.restassured.RestAssured.given;
 
+@Ignore("legacy test that should be converted or deleted")
 @RequiresVersion(">=1.1.0")
 @RequiresAuthentication
 public class StatsTest extends BaseRestTest {
@@ -33,22 +34,22 @@ public class StatsTest extends BaseRestTest {
     public void testSystemStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix)
+                .get(resourcePrefix)
                 .then()
-                    .statusCode(200)
-                    .assertThat()
-                        .body(".", containsAllKeys("fs", "jvm", "network", "os", "process"));
+                .statusCode(200)
+                .assertThat()
+                .body(".", containsAllKeys("fs", "jvm", "network", "os", "process"));
     }
 
     @Test
     public void testFsStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix + "/fs")
+                .get(resourcePrefix + "/fs")
                 .then()
-                    .statusCode(200)
-                    .assertThat()
-                        .body(".", containsAllKeys("filesystems"));
+                .statusCode(200)
+                .assertThat()
+                .body(".", containsAllKeys("filesystems"));
 
     }
 
@@ -56,35 +57,35 @@ public class StatsTest extends BaseRestTest {
     public void testJvmStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix + "/jvm")
+                .get(resourcePrefix + "/jvm")
                 .then()
-                    .statusCode(200);
+                .statusCode(200);
     }
 
     @Test
     public void testNetworkStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix + "/network")
+                .get(resourcePrefix + "/network")
                 .then()
-                    .statusCode(200);
+                .statusCode(200);
     }
 
     @Test
     public void testOsStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix + "/os")
+                .get(resourcePrefix + "/os")
                 .then()
-                    .statusCode(200);
+                .statusCode(200);
     }
 
     @Test
     public void testProcessStats() throws Exception {
         given()
                 .when()
-                    .get(resourcePrefix + "/process")
+                .get(resourcePrefix + "/process")
                 .then()
-                    .statusCode(200);
+                .statusCode(200);
     }
 }
