@@ -17,18 +17,20 @@
 package integration.system.bundles;
 
 import com.google.common.net.HttpHeaders;
-import com.jayway.restassured.response.ValidatableResponse;
 import integration.BaseRestTest;
 import integration.RequiresAuthentication;
 import integration.RequiresVersion;
+import io.restassured.response.ValidatableResponse;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.net.URI;
 
-import static com.jayway.restassured.RestAssured.given;
+import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
+@Ignore("legacy test that should be converted or deleted")
 @RequiresVersion(">=1.2.0")
 @RequiresAuthentication
 public class ContentPackStreamsTest extends BaseRestTest {
