@@ -20,11 +20,13 @@ import integration.BaseRestTest;
 import integration.MongoDbSeed;
 import integration.RequiresAuthentication;
 import integration.RequiresVersion;
+import org.junit.Ignore;
 import org.junit.Test;
 
-import static com.jayway.restassured.RestAssured.given;
+import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.notNullValue;
 
+@Ignore("legacy test that should be converted or deleted")
 @RequiresVersion(">=0.90.0")
 @RequiresAuthentication
 public class InputsTest extends BaseRestTest {
@@ -36,7 +38,7 @@ public class InputsTest extends BaseRestTest {
         given().when()
                 .body(jsonResourceForMethod()).post("/system/inputs")
                 .then()
-                    .statusCode(400).statusLine(notNullValue());
+                .statusCode(400).statusLine(notNullValue());
     }
 
     @Test
